facade-service/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
requests==2.22.0
uvicorn==0.17.6
kafka-python==2.0.2
python-consul==1.1.0